requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-DotEnv==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.0.2
